backend/requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.12.0
pytest-django==4.12.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0